+++ go-server/handler/warehouse/addWarehouse.go
@@ -1,4 +1,4 @@
-package handler
+package warehouse
 
 import (
 	"Go_simpleWMS/utils"
@@ -77,32 +77,4 @@ func AddWarehouse(context *gin.Context) {
 	}
 
 	context.JSON(http.StatusOK, gin.H{"message": "Warehouse added successfully"})
-}
-
-func DeleteWarehouse(context *gin.Context) {
-	wid := context.PostForm("wid")
-	if wid == "" {
-		context.JSON(http.StatusBadRequest, gin.H{"message": "wid is required"})
-		return
-	}
-
-	tx, _ := utils.GetDbConnection()
-
-	if tx == nil {
-		context.JSON(http.StatusInternalServerError, gin.H{"error": "Cannot begin transaction"})
-		return
-	}
-
-	// 删除仓库
-	_, err := tx.Exec("DELETE FROM warehouse WHERE wid=?", wid)
-	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{"error": "Cannot delete the warehouse"})
-		return
-	}
-	err = tx.Commit()
-	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{"error": "Cannot commit the transaction"})
-		return
-	}
-	context.JSON(http.StatusOK, gin.H{"message": "Warehouse deleted successfully"})
 }
